refactor: controller with decorator
